Split into two files - generic tensor operations and classifications
